--- a/Generator/Sources/NeedleFramework/Generating/Pluginized/PluginizedDependencyProviderSerializerTask.swift
+++ b/Generator/Sources/NeedleFramework/Generating/Pluginized/PluginizedDependencyProviderSerializerTask.swift
@@ -94,7 +94,7 @@ class PluginizedDependencyProviderSerializerTask: AbstractTask<[SerializedProvid
             return ProviderAttributes()
         }
         var maxLevel: Int = 0
-        provider.levelMap.forEach { (componentType: String, level: Int) in
+        for (_, level) in provider.levelMap {
             if level > maxLevel {
                 maxLevel = level
             }