… Generator/Sources/NeedleFramework/Generating/Serializers/DependencyProviderSerializer.swift
… Generator/Tests/NeedleFrameworkTests/Generating/DependencyGraphExporterTests.swift
… Generator/Tests/NeedleFrameworkTests/Generating/DependencyProviderSerializerTaskTests.swift
… Generator/Tests/NeedleFrameworkTests/Generating/Pluginized/PluginizedDependencyGraphExporterTests.swift
… Sample/MVC/TicTacToe/Sources/NeedleGenerated.swift
… Sample/Pluginized/TicTacToe/TicTacToeCore/NeedleGenerated.swift